chore: 각도, 이동거리, 장애물 거리 int로 변경

+++ analysis-service/src/main/kotlin/com/parkit/analysis/coaching/dto/CoachingSocketDto.kt
@@ -13,10 +13,11 @@ data class ObstacleDistancesDto(
 data class CoachingSocketDto(
     val step: Int,
     val timestamp: Long,
-    val targetAngle: Double,
-    val targetDistance: Double,
-    val currentAngle: Double,
-    val currentDistance: Double,
+    // units: angle=deg, distance=cm
+    val targetAngle: Int,
+    val targetDistance: Int,
+    val currentAngle: Int,
+    val currentDistance: Int,
 	val distances: ObstacleDistancesDto,
     val coachingId: Int,
 )
